Switch to manifest_version 3, drop github permission
Upgrade to manifest_version 3, which is mostly the same, except we now
need to request an additional "webRequestFilterResponse" permission and
split out the host permissions.

Persistent background scripts are no longer allowed, so drop it. I
reviewed the documentation and we seem to already be following all the
correct practices by registering listeners at the top level. Also in
testing, I stopped the background script and it resumed as desired.

Because v3 has an optional_host_permissions, I realized the
api.github.com permission isn't actually required, presumably because
we're not accessing any data from it, just sending off a request.